Fix non-running conditional autocmd of ColorScheme
Even though vimrc adds the conditional ColorScheme autocmd and then
later forces its running with doautocmd in vimrc, Nvim doesn't want to
run those autocmd during Nvim initialization for unknown reason. (but
they run fine afterwards when Nvim is running, and it was fine in the
past when the line was just added). Somehow putting if compound
statements in a function and calling it solves the problem.